// src/app/Type.ts
export type BazarProduct = {
  id: number
  slug: string
  nameBn: string
  category: string
  categoryNameBn: string
  categoryIcon: string
  unit: string
  image: string
  today: number
  yesterday: number
  lastWeek: number
  lastMonth: number
  change: {
    dir: 'up' | 'down'
    pct: number
  }
}